backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
Pillow==10.2.0
gunicorn==21.2.0
requests==2.31.0
google-auth==2.27.0
